# Conflicts:
#	.github/.build-env

+++ Jenkinsfile
@@ -48,10 +48,19 @@ pipeline {
 
         stage('Prepare the build') {
             steps {
-                writeFile file: 'changelog.txt', text: "\n= ${params.buildVersion} - ${releaseDate} =\n"+params.changelog+"\n"
-                writeFile file: 'changelog.md', text: "\n### ${params.buildVersion} - ${releaseDate}\n"+params.changelog+"\n"
-                sh "./jenkins/version-update.sh ${params.buildVersion} ${params.minProVersion} ${params.editorVersion} ${params.syncVersion} ./changelog.txt ./changelog.md"
-                sh "rm -rf changelog.txt changelog.md"
+                sh "./jenkins/version-update.sh ${params.buildVersion} ${params.minProVersion} ${params.editorVersion} ${params.syncVersion}"
+            }
+        }
+
+        stage('Prepare changelog') {
+            when {
+                expression { return params.svnCommit }
+            }
+            steps {
+                 writeFile file: 'changelog.txt', text: "\n= ${params.buildVersion} - ${releaseDate} =\n"+params.changelog+"\n"
+                 writeFile file: 'changelog.md', text: "\n### ${params.buildVersion} - ${releaseDate}\n"+params.changelog+"\n"
+                 sh "./jenkins/changelog-update.sh ./changelog.txt ./changelog.md"
+                 sh "rm -rf changelog.txt changelog.md"
             }
         }
